// angular-components-oasis/projects/oasis-button-expand/src/shared/button-expand-state.enum.ts
/**
 * Состояние кнопки
 */
export enum ButtonExpandState {
  /** Стандартное состояние */
  Default = 'default',

  /** Наведённое состояние */
  Hover = 'hover',

  /** Нажатое состояние */
  Click = 'click',

  /** Выключенное состояние */
  Disabled = 'disabled'
}