// angular-components-oasis/projects/oasis-button-expand/src/shared/button-expand-unit.enum.ts
/**
 * Основные типы данных для установки размера компонента
 */
export enum ButtonExpandUnit {
  /** Не выбран тип */
  None = '',

  /** Пиксели */
  Pixel = 'px',

  /** Проценты */
  Percent = '%',

  /** Ширина экрана */
  ViewportWidth = 'vw',

  /** Высота экрана */
  ViewportHeight = 'vh'
}